+++ theme/js/components/search/search.vue
@@ -135,9 +135,7 @@
           </div>
           <ul v-else-if="results.length" class="fr-mt-1w border-default-grey border-top">
             <li v-for="result in results" :key="result.id">
-              <a :href="result.page" class="unstyled fr-raw-link w-10 block">
-                <Dataset v-bind="result" />
-              </a>
+              <Dataset v-bind="result" />
             </li>
             <Pagination
               v-if="totalResults > pageSize"
@@ -229,6 +227,7 @@ export default defineComponent({
 
     /**
      * Search results
+     * @type {Ref<Array>}
      */
     const results = ref([]);
 
@@ -276,6 +275,18 @@ export default defineComponent({
     const searchRef = ref(null);
 
     /**
+     * 
+     * @param {Array} data 
+     */
+    const formatResults = (data) => {
+      results.value = data.map(result => {
+        result.last_modified = new Date(result.last_modified);
+        return result;
+      });
+      return results;
+    };
+
+    /**
      * Search new dataset results
      */
     const search = () => {
@@ -292,7 +303,7 @@ export default defineComponent({
         })
         .then((res) => res.data)
         .then((result) => {
-          results.value = result.data;
+          formatResults(result.data);
           totalResults.value = result.total;
           loading.value = false;
         })
@@ -445,7 +456,7 @@ export default defineComponent({
         if (total && parseInt(total) > 0) {
           let datasetResults = resultsRef.value.dataset.results;
           if(datasetResults) {
-            results.value = JSON.parse(datasetResults);
+            formatResults(JSON.parse(datasetResults));
           }
           totalResults.value = JSON.parse(total);
         }