+++ theme/js/components/utils/placeholder.vue
@@ -8,17 +8,18 @@ category: 5 - Interactions
 
 A very simple component that takes a `type` (dataset, reuse, post...) and an optional `src`. If `src` is defined, it simply displays the image (and all props passed are passed to the image).
 However, if `src` is undefined, it falls back to a placeholder. Don't forget to pass the `alt` prop !
-Sometimes you don't need an image but a `<div>` with a background-image property. Simply pass the `backgroundImage` prop and you will get a nice div.
 
 -->
 
 <template>
-  <div
-    v-if="backgroundImage"
+  <img
+    loading="lazy"
+    :src="path"
+    :alt="alternativeTextForDefinedImageOnly"
+    :width="size"
+    :height="size"
     v-bind="$attrs"
-    :style="{ backgroundImage: `url('${path}')` }"
   />
-  <img loading="lazy" :src="path" :alt="alternativeTextForDefinedImageOnly" v-bind="$attrs" v-else />
 </template>
 
 <script>
@@ -32,7 +33,7 @@ export default {
     type: String,
     src: String,
     alt: String,
-    backgroundImage: Boolean,
+    size: Number,
   },
   computed: {
     alternativeTextForDefinedImageOnly() {