+++ theme/js/composables/useExternalSource.js
@@ -1,0 +1,8 @@
+/**
+ * 
+ * @param {Object} extras 
+ * @returns {String | undefined}
+ */
+export default function useExternalSource(extras) {
+    return extras?.remote_url;
+}